동적 라이브러리 링킹으로 인한 워닝 해결 (프로젝트별 vcpkg 설정에서 static 링킹하도록)

+++ JunCore/JunCore/network/IOCPManager.cpp
@@ -92,7 +92,7 @@ void IOCPManager::HandleRecvComplete(Session* session, DWORD ioSize)
 		}
 
 		// 충분한 패킷 데이터가 도착했는지 확인
-        if (_recv_byte < _packet_len)
+        if (static_cast<uint32_t>(_recv_byte) < _packet_len)
         {
             break;
         }
